feat(web): add a floating toolbar
