+++ apps/web/src/components/issues/issue-item.tsx
@@ -3,14 +3,8 @@ import Link from 'next/link'
 import type { TIssue } from '@buildit/utils/types'
 
 import { Avatar, AvatarFallback, AvatarImage } from '@buildit/ui/avatar'
-import {
-  ContextMenu,
-  ContextMenuContent,
-  ContextMenuGroup,
-  ContextMenuItem,
-  ContextMenuShortcut,
-  ContextMenuTrigger,
-} from '@buildit/ui/context-menu'
+import { Checkbox } from '@buildit/ui/checkbox'
+import { cn } from '@buildit/ui/cn'
 
 import { Icons } from '@/components/ui/icons'
 import { priorities, statuses } from '@/configs/issue-types'
@@ -29,16 +23,15 @@ type IssueItemProps = Pick<
   | 'project'
 >
 
-/**
- * The Issue Item component to display a single issue.
- * @param props The component properties.
- * @param props.issue The issue to display.
- * @returns JSX component.
- */
-export default function IssueItem({ issue }: { issue: IssueItemProps }) {
-  const updatedAt = issue.updatedAt ? formatDate(issue.updatedAt) : undefined
-  const createdAt = issue.createdAt ? formatDate(issue.createdAt) : undefined
-
+const IssueItem = ({
+  issue,
+  isFirst,
+  isLast,
+}: {
+  issue: IssueItemProps
+  isFirst: boolean
+  isLast: boolean
+}) => {
   const priorityIconName = priorities.find(
     (priority) => priority.value === issue.priority,
   )?.icon
@@ -49,49 +42,54 @@ export default function IssueItem({ issue }: { issue: IssueItemProps }) {
   const PriorityIcon = Icons[priorityIconName as keyof typeof Icons]
   const StatusIcon = Icons[statusIconName as keyof typeof Icons]
 
+  const updatedAt = issue.updatedAt ? formatDate(issue.updatedAt) : undefined
+  const createdAt = issue.createdAt ? formatDate(issue.createdAt) : undefined
+
   return (
-    <>
-      <ContextMenu>
-        <ContextMenuTrigger>
-          <Link
-            href={`/issue/${issue.issueId}`}
-            className='flex items-center justify-between border-b px-5 py-2 hover:bg-gray-100/50'
-          >
-            <div className='flex items-center space-x-4'>
-              <PriorityIcon className='size-4 text-sub' />
-              <p className='text-sm text-sub'>{issue.issueId}</p>
-              <StatusIcon className='size-4 text-sub' />
-              <p className='text-sm text-surface'>{issue.title}</p>
-            </div>
-            <div className='flex items-center gap-x-3'>
-              <span className='text-soft text-xs'>{updatedAt}</span>
-              <span className='text-soft text-xs'>{createdAt}</span>
-              {issue.assigneeId && issue.assignee ? (
-                <Avatar className='size-5'>
-                  <AvatarImage
-                    src={issue.assignee.image ?? ''}
-                    alt={issue.assignee.name ?? ''}
-                  />
-                  <AvatarFallback>
-                    {issue.assignee.name?.charAt(0)}
-                  </AvatarFallback>
-                </Avatar>
-              ) : (
-                <div>hello</div>
-              )}
-            </div>
-          </Link>
-        </ContextMenuTrigger>
-        <ContextMenuContent className='w-56'>
-          <ContextMenuGroup>
-            <ContextMenuItem>
-              <Icons.trash className='mr-2 h-4 w-4' />
-              Delete
-              <ContextMenuShortcut>Delete</ContextMenuShortcut>
-            </ContextMenuItem>
-          </ContextMenuGroup>
-        </ContextMenuContent>
-      </ContextMenu>
-    </>
+    <Link href={`/issue/${issue.issueId}`}>
+      <div
+        className={cn(
+          ' grid grid-cols-[500px_2fr] border-x p-3 bg-white hover:bg-weak/50 transition-colors duration-200',
+          isFirst && 'rounded-t-lg border-t',
+          isLast ? 'rounded-b-lg border-b' : 'border-b',
+        )}
+      >
+        <div className='grid grid-cols-[minmax(130px,_150px)_2fr] gap-2 items-center'>
+          <div className='group grid grid-cols-[20px_20px_1fr_20px] gap-2 items-center'>
+            <button
+              onClick={(e) => {
+                e.stopPropagation()
+                e.preventDefault()
+              }}
+              role='checkbox'
+              aria-checked='false'
+              className='invisible group-hover:visible'
+            >
+              <Checkbox className='data-[state=checked]:visible' />
+            </button>
+
+            <PriorityIcon className='size-4 text-sub' />
+            <p className='text-sm text-soft'>{issue.issueId}</p>
+            <StatusIcon className='size-4 text-sub' />
+          </div>
+          <p className='text-sub text-sm font-medium'>{issue.title}</p>
+        </div>
+        <div className='flex items-center justify-end gap-3'>
+          <span className='text-soft text-xs'>{updatedAt}</span>
+          <span className='text-soft text-xs'>{createdAt}</span>
+          {issue.assigneeId && issue.assignee && (
+            <Avatar className='size-5'>
+              <AvatarImage
+                src={issue.assignee.image ?? ''}
+                alt={issue.assignee.name ?? ''}
+              />
+              <AvatarFallback>{issue.assignee.name?.charAt(0)}</AvatarFallback>
+            </Avatar>
+          )}
+        </div>
+      </div>
+    </Link>
   )
 }
+
+export default IssueItem
